+++ createorto.h
@@ -38,7 +38,7 @@ public slots:
     void cambioestadoComboBox(int estado );
     void cambioestadoLineEdit(QString directorio);
 
-    void VigilarTamanyPixel(int tamanyoPixel);
+    void VigilarTamanyPixel(double tamanyoPixel);
     void VigilarUtm(int utm);
     void VigilarAnchoPasada(int anchoPasada);
     void VigilarOffsetPasada(int offsetPasada);
